+++ backend/src/main/java/com/stmanagement/model/SupplierOrderDetail.java
@@ -1,7 +1,6 @@
 package com.stmanagement.model;
 import lombok.AllArgsConstructor; import lombok.Builder; import lombok.Data; import lombok.NoArgsConstructor;
 import javax.persistence.*;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data @NoArgsConstructor @AllArgsConstructor @Builder
@@ -11,9 +10,9 @@ public class SupplierOrderDetail {
     @Column(name = "order_id", nullable = false) private Long orderId;
     @Column(name = "employee_name", length = 100) private String employeeName;
     @Column(name = "item_name", length = 500, nullable = false) private String itemName;
-    @Column(nullable = false) private BigDecimal quantity;
-    @Column(name = "unit_price", nullable = false) private BigDecimal unitPrice;
-    @Column(nullable = false) private BigDecimal amount;
+    @Column(nullable = false) private Double quantity;
+    @Column(name = "unit_price", nullable = false) private Double unitPrice;
+    @Column(nullable = false) private Double amount;
     @Column(length = 500) private String remark;
     @Column(name = "create_time", nullable = false, updatable = false) private LocalDateTime createTime;
     @PrePersist protected void onCreate() { createTime = LocalDateTime.now(); }
